Cleanup `index.js` exports
- only export public functions
- remove use of `this`
- move helper functions to a module
- move `removeTagsFromBuffer` up
