+++ src/ID3Util.js
